Refactor only: move WAL stuff into its own files
