+++ vindex/map_test.go
@@ -21,12 +21,9 @@ import (
 	"context"
 	"crypto/sha256"
 	"encoding/hex"
-	"fmt"
-	"io"
 	"iter"
 	"os"
 	"testing"
-	"time"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/transparency-dev/formats/log"
@@ -34,248 +31,12 @@ import (
 	"github.com/transparency-dev/merkle/rfc6962"
 	"github.com/transparency-dev/merkle/testonly"
 	"golang.org/x/mod/sumdb/note"
-	"golang.org/x/sync/errgroup"
 )
 
 const (
 	skey = "PRIVATE+KEY+logandmap+38581672+AXJ0FKWOcO2ch6WC8kP705Ed3Gxu7pVtZLhfHAQwp+FE"
 	vkey = "logandmap+38581672+Ab/PCr1eCclRPRMBqw/r5An1xO71MCnImLiospEq6b4l"
 )
-
-func TestWriteAheadLog_init(t *testing.T) {
-	testCases := []struct {
-		desc         string
-		fileContents string
-		wantIdx      uint64
-		wantErr      bool
-	}{
-		{
-			desc:         "empty file",
-			fileContents: "",
-			wantIdx:      0,
-			wantErr:      false,
-		}, {
-			desc:         "0 file",
-			fileContents: "0\n",
-			wantIdx:      1,
-			wantErr:      false,
-		}, {
-			desc:         "just indexes",
-			fileContents: "0\n1\n2\n",
-			wantIdx:      3,
-			wantErr:      false,
-		}, {
-			desc:         "indexes and hashes",
-			fileContents: fmt.Sprintf("1 %s %s\n", mustHashEncode("1a"), mustHashEncode("1b")),
-			wantIdx:      2,
-			wantErr:      false,
-		}, {
-			desc:         "trailing corruption",
-			fileContents: "1\n2 fdfxx",
-			wantErr:      true,
-		}, {
-			desc:         "lots of newlines",
-			fileContents: "1\n2\n3\n\n",
-			wantErr:      true,
-		}, {
-			desc:         "no trailing newlines",
-			fileContents: "1\n2\n3",
-			wantErr:      true,
-		},
-	}
-	for _, tC := range testCases {
-		t.Run(tC.desc, func(t *testing.T) {
-			f, err := os.CreateTemp("", "testWal")
-			if err != nil {
-				t.Fatal(err)
-			}
-			if _, err := f.WriteString(tC.fileContents); err != nil {
-				t.Fatal(err)
-			}
-			if err := f.Close(); err != nil {
-				t.Fatal(err)
-			}
-			wal := &walWriter{
-				walPath: f.Name(),
-			}
-			idx, err := wal.init()
-			if gotErr := err != nil; gotErr != tC.wantErr {
-				t.Fatalf("wantErr != gotErr (%t != %t) %v", tC.wantErr, gotErr, err)
-			}
-			defer func() {
-				_ = wal.close()
-			}()
-			if tC.wantErr {
-				return
-			}
-			if idx != tC.wantIdx {
-				t.Errorf("want idx %v but got %v", tC.wantIdx, idx)
-			}
-		})
-	}
-}
-
-func TestWriteAheadLog_roundtrip(t *testing.T) {
-	f, err := os.CreateTemp("", "testWal")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := f.Close(); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.Remove(f.Name()); err != nil {
-		t.Fatal(err)
-	}
-
-	wal := &walWriter{
-		walPath: f.Name(),
-	}
-	idx, err := wal.init()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got, want := idx, uint64(0); got != want {
-		t.Fatalf("expected index %d, got %d", want, got)
-	}
-
-	for i := range 33 {
-		hash := sha256.Sum256([]byte{byte(i)})
-		if err := wal.append(uint64(i), [][32]byte{hash}); err != nil {
-			t.Error(err)
-		}
-	}
-
-	if err := wal.close(); err != nil {
-		t.Error(err)
-	}
-
-	idx, err = wal.init()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got, want := idx, uint64(33); got != want {
-		t.Fatalf("expected index %d, got %d", want, got)
-	}
-
-	if err := wal.close(); err != nil {
-		t.Error(err)
-	}
-}
-
-func TestWriteAndWriteLog(t *testing.T) {
-	f, err := os.CreateTemp("", "testWal")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := f.Close(); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.Remove(f.Name()); err != nil {
-		t.Fatal(err)
-	}
-
-	wal := &walWriter{
-		walPath: f.Name(),
-	}
-	idx, err := wal.init()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got, want := idx, uint64(0); got != want {
-		t.Fatalf("expected index %d, got %d", want, got)
-	}
-
-	reader, err := newWalReader(f.Name())
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	const count = 2056
-	var eg errgroup.Group
-	eg.Go(func() error {
-		for i := range count {
-			hash := sha256.Sum256([]byte{byte(i)})
-			if err := wal.append(uint64(i), [][32]byte{hash}); err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-	eg.Go(func() error {
-		var expect uint64
-		for expect < count {
-			idx, _, err := reader.next()
-			if err != nil {
-				if err != io.EOF {
-					return err
-				}
-				// Wait a small amount of time for more data to become available
-				time.Sleep(10 * time.Millisecond)
-				continue
-			}
-			if got, want := idx, expect; got != want {
-				return fmt.Errorf("expected index %d, got %d", want, got)
-			}
-			expect++
-		}
-		return nil
-	})
-	if err := eg.Wait(); err != nil {
-		t.Fatal(err)
-	}
-
-	if err := wal.close(); err != nil {
-		t.Error(err)
-	}
-	if err := reader.close(); err != nil {
-		t.Error(err)
-	}
-}
-
-func TestUnmarshal(t *testing.T) {
-	testCases := []struct {
-		desc       string
-		entry      string
-		wantErr    bool
-		wantIdx    uint64
-		wantHashes int
-	}{
-		{
-			desc:       "just index",
-			entry:      "1",
-			wantErr:    false,
-			wantIdx:    1,
-			wantHashes: 0,
-		}, {
-			desc:       "index and hashes",
-			entry:      fmt.Sprintf("1 %s %s", mustHashEncode("1a"), mustHashEncode("1b")),
-			wantErr:    false,
-			wantIdx:    1,
-			wantHashes: 2,
-		}, {
-			desc:    "corruption at the end",
-			entry:   "1 deadbeef feed01xxx",
-			wantErr: true,
-		},
-	}
-	for _, tC := range testCases {
-		t.Run(tC.desc, func(t *testing.T) {
-			idx, hashes, err := unmarshalWalEntry(tC.entry)
-			if gotErr := err != nil; gotErr != tC.wantErr {
-				t.Fatalf("wantErr != gotErr (%t != %t) %v", tC.wantErr, gotErr, err)
-			}
-			if tC.wantErr {
-				return
-			}
-			if idx != tC.wantIdx {
-				t.Errorf("want idx %v but got %v", tC.wantIdx, idx)
-			}
-			if got, want := len(hashes), tC.wantHashes; got != want {
-				t.Errorf("want %v hashes but got %v: %q", want, got, hashes)
-			}
-		})
-	}
-}
 
 func TestVerifiableIndex(t *testing.T) {
 	ctx := context.Background()
